==> scripts/01_ingest.R <==
suppressPackageStartupMessages({ library(tidyverse); library(readxl); library(yaml); source("R/functions/io.R") })
cfg <- cfg_read(); ddir <- cfg$data_dir; odir <- cfg$outputs_dir; ensure_dir(odir); ensure_dir(file.path(odir,"figures")); ensure_dir(file.path(odir,"tables"))
load(file.path(ddir, cfg$files$metabolon_rdata))
load(file.path(ddir, cfg$files$metabolomics_index))
form1 <- read_csv(file.path(ddir, cfg$files$form1), show_col_types = FALSE)
form4 <- read_csv(file.path(ddir, cfg$files$form4), show_col_types = FALSE)
form5 <- read_csv(file.path(ddir, cfg$files$form5), show_col_types = FALSE)
form9 <- read_csv(file.path(ddir, cfg$files$form9), show_col_types = FALSE)
markers <- read_csv(file.path(ddir, cfg$files$form_markers), show_col_types = FALSE)
dental <- readxl::read_excel(file.path(ddir, cfg$files$dental_events_xlsx))
cimt_cac <- tryCatch(read_csv(file.path(ddir, cfg$files$cimt_cac_csv), show_col_types = FALSE), error = function(e) NULL)
save(form1, form4, form5, form9, markers, dental, cimt_cac, file = file.path(odir, "01_forms.RData"))
save.image(file = file.path(odir, "01_workspace_raw.RData"))
