==> scripts/05_latent_factors.R <==
suppressPackageStartupMessages({ library(tidyverse); library(sva); library(bcv); library(cate) })
load("outputs/03_dat_with_phenotypes.RData")
hdl_col <- which(colnames(dat) == "LAB_HDL_VAP"); meta_cols <- 13:(hdl_col - 1); metab <- as.matrix(dat[, meta_cols])
mod <- model.matrix(~ Gingivitis + AGE + GENDER + RACE2 + PE_BP_CLASS + PE_SYSTOLIC + SCR_CURSMOKE + DEMO_DIABETES + HS_hsCRP.log + HS_IL6.log + LAB_CHOL_VAP + LAB_HDL_VAP, data = dat)
k_be <- sva::num.sv(t(metab), mod, method = "be"); k_leek <- sva::num.sv(t(metab), mod, method = "leek"); k_bcv <- tryCatch(bcv::bcv(t(metab))$best, error = function(e) NA_integer_)
k_default <- 20L; k <- ifelse(is.na(k_bcv), max(k_be, k_leek, k_default), max(k_be, k_leek, k_bcv, k_default))
fit_cate <- cate::cate(~ Gingivitis | AGE + GENDER + RACE2 + PE_BP_CLASS + PE_SYSTOLIC + SCR_CURSMOKE + DEMO_DIABETES + HS_hsCRP.log + HS_IL6.log + LAB_CHOL_VAP + LAB_HDL_VAP, X.data = dat, Y = metab, r = k)
C_hat <- fit_cate$Z; saveRDS(C_hat, "outputs/05_latent_factors.rds")
latent_summary <- tibble(method = c("be", "leek", "bcv", "chosen"), k_est = c(k_be, k_leek, k_bcv, ncol(C_hat))); write.csv(latent_summary, "outputs/tables/05_latent_summary.csv", row.names = FALSE)
